pyproject.toml
[project]
name = "documentation-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "certifi>=2024.0.0",
    "langchain>=1.0.3",
    "langchain-chroma>=0.1.0",
    "langchain-classic>=0.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=0.1.0",
    "langchain-openai>=1.0.2",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.13",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
